Remove LSM Reader and LSM Toolbox from fiji.git
It was the original hope by this maintainer that the developers of
aforementioned plugins would jump at the opportunity to optimize their
development workflow by integrating it into the Fiji workflow.

However, that did not happen, and unfortunately problems arose because
of it: the LSM support of LSM Reader/Toolbox still lacks features that
Bio-Formats addressed long ago, and Bio-Formats still lacks the
performance of LSM Reader/Toolbox. In short: the synergy that would have
been the consequence of an efficient, unified workflow, never happened.

The silver lining is that the SCIFIO project (a re-architecturing of
Bio-Formats) promises to address both concerns, so it is time to
deprecate the LSM Reader/Toolbox in Fiji.

It is still available from the IBMP-CNRS update site listed in
http://fiji.sc/List_of_update_sites, so people who relied on the LSM
Reader/Toolbox functionality are not *completely* left behind ;-)

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>