TrackMate_: fix the semi-auto tracking.
With recent the changes in TrackMate, we had lost the ability to pass a
calibrated (as in having pixel sizes) neighborhood to the detector. This
is needed however, since the detector needs to know whether the source
image has an anisotropic calibration.
We now ask the spot neighborhood to ship with a calibration array. The
transform it shipped is still there, but is use only to map the found
spot coordinates to a global coordinate system, since the detector
already returns spots in the source image coordinates.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>