MaMuT_: performance improvement of the track overlay.
The track drawing is the slowest part of MaMuT, and is punitive when a
large number of large tracks are to be drawn. This commit improves the
drawing performance by using the tweaks introduced recently in
TrackMate.

Unfortunately, the track rendering is still too slow. However now, the
slow part is cornered to be the call to native drawing method
sun.java2d.SunGraphics2D.drawLine(int, int, int, int)

The Yourkit profiler (v12.0.2) reports that this method takes 99% of the
time of
net.imglib2.ui.InteractiveDisplayCanvasComponent.paintComponent(Graphics).
Event fine profiling concord.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>